Move observer to metrics package
