+++ pkg/controller/controller.go
@@ -42,7 +42,7 @@ type Controller struct {
 	canaries      *sync.Map
 	jobs          map[string]CanaryJob
 	deployer      CanaryDeployer
-	observer      CanaryObserver
+	observer      metrics.CanaryObserver
 	recorder      metrics.CanaryRecorder
 	notifier      *notifier.Slack
 	meshProvider  string
@@ -81,10 +81,6 @@ func NewController(
 		},
 	}
 
-	observer := CanaryObserver{
-		metricsServer: metricServer,
-	}
-
 	recorder := metrics.NewCanaryRecorder(controllerAgentName, true)
 	recorder.SetInfo(version, meshProvider)
 
@@ -101,7 +97,7 @@ func NewController(
 		jobs:          map[string]CanaryJob{},
 		flaggerWindow: flaggerWindow,
 		deployer:      deployer,
-		observer:      observer,
+		observer:      metrics.NewObserver(metricServer),
 		recorder:      recorder,
 		notifier:      notifier,
 		meshProvider:  meshProvider,
